Moved api.js ans maintenance.js one level deeper to api directory
