+++ apis/server/api.js
@@ -11,7 +11,7 @@ import { Roles } from 'meteor/alanning:roles';
 
 // Collection imports
 import Apis from '/apis/collection';
-import ApiV1 from '/core/server/api';
+import ApiV1 from '/core/server/api/api';
 import Organizations from '/organizations/collection';
 
 ApiV1.swagger.meta.paths = {
