+++ organizations/server/api.js
@@ -7,7 +7,7 @@
 import { Meteor } from 'meteor/meteor';
 
 // Collection imports
-import MaintenanceV1 from '/core/server/maintenance';
+import MaintenanceV1 from '/core/server/api/maintenance';
 import Organizations from '/organizations/collection';
 import { Accounts } from 'meteor/accounts-base';
